+++ src/base.py
@@ -1,86 +1,48 @@
 # Standard Library Imports
 import os
-import pickle
 from timeit import default_timer as timer
 # Third Party Imports
+import numpy as np
 import pandas as pd
 # Local Application Imports
-from src.data.make_dataset import table_cols_info as cols_info
-from src.features.build_features import train_test_split
-from src.models.train_model import train_model, test_model
-
-# Global Variables
-DATA_RAW = os.environ['DATA_RAW']
-DATA_INTERIM = os.environ['DATA_INTERIM']
-DATA_PROCESSED = os.environ['DATA_PROCESSED']
-APP_MODELS = os.environ['APP_MODELS']
-APP_REPORTS = os.environ['APP_REPORTS']
-APP_VERSION = os.environ['APP_VERSION']
-
-
-def _save_model(clf, filename=None):
-    if filename is None:
-        filename = APP_VERSION + '_decision_tree.sav'
-    pickle.dump(clf, open(APP_MODELS + filename, 'wb'))
-
-
-def _load_model(filename=None):
-    if filename is None:
-        filename = APP_VERSION + '_decision_tree.sav'
-    clf = pickle.load(open(APP_MODELS + filename, 'rb'))
-
-    return clf
-
-
-def _split_raw():
-    # Retrieve data from raw folder (assume it's already masked and cleaned data)
-    raw_orders = pd.read_csv(DATA_RAW + 'orders.csv',
-                             parse_dates=[cols_info['registry_date']],
-                             index_col=0)
-
-    # Get data in a proper format
-    x_train, x_test, y_train, y_test = train_test_split(raw_orders)
-
-    return x_train, x_test, y_train, y_test
-
-
-def _save_processed(x_train, x_test, y_train, y_test):
-    # Save train and test sets
-    x_train.merge(y_train.to_frame(), left_index=True, right_index=True)\
-        .to_csv(DATA_PROCESSED + 'train.csv')
-    x_test.merge(y_test.to_frame(), left_index=True, right_index=True)\
-        .to_csv(DATA_PROCESSED + 'test.csv')
-
-
-def _load_processed():
-
-    df_test = pd.read_csv(DATA_PROCESSED + 'test.csv', index_col=0)
-    y_test = df_test['target']
-    x_test = df_test.drop(columns='target')
-
-    df_train = pd.read_csv(DATA_PROCESSED + 'train.csv', index_col=0)
-    y_train = df_train['target']
-    x_train = df_train.drop(columns='target')
-
-    return x_train, x_test, y_train, y_test
+from src.data.make_dataset import make_dataset
+from src.features.build_features import transform_raw, load_processed, save_processed
+from src.models.train_model import train_model, test_model, load_model, save_model
+from src.visualization.visualize import plot_learning_curve, plot_confusion_matrix, plot_roc_curve, plot_tree
+from src.visualization.visualize import plot_precision_recall_curve
 
 
 if __name__ == "__main__":
 
+    table_names = ['orders']
     model = None
     x_train = pd.DataFrame()
     x_test = pd.DataFrame()
     y_train = pd.DataFrame()
     y_test = pd.DataFrame()
 
+    # Maximum number of nodes allowed for exporting the tree structure
+    nodes_warning = 100
+    # Maximum number of columns allowed to generate a pandas_profiling
+    report_warning = 100
+
+    # Features to be built
+    feats = ['index', 'recency_v0', 'recency_v1', 'c_age', 'gravity', 'monetary', 'frequency', 'tsfresh', 'target']
+    scoring = 'roc_auc'
+
+    # Number of cross-validations folds
+    cv = 5
+    # Prints title in figures.
+    title = False
+
     # Set grid search params
     grid_params = [{
-        'clf__criterion': ['gini', 'entropy'],
-        'clf__class_weight': ['balanced'],
+        # 'clf__criterion': ['gini', 'entropy'],
+        # 'clf__class_weight': ['balanced', None],
+        'clf__min_samples_split': np.arange(50, 200, 20),
+        'clf__min_impurity_decrease': np.linspace(0.0, 0.002, 5),
+        'clf__max_depth': np.arange(3, 6, 1),
         'clf__random_state': [42]
-        # 'clf__min_samples_leaf': np.linspace(0.001, 0.05, 5),
-        # 'clf__max_depth': np.arange(1, 10, 1),
-        # 'clf__min_samples_split': np.linspace(0.0001, 0.050, 5)
     }]
 
     print("Running script...\n")
@@ -93,54 +55,96 @@ if __name__ == "__main__":
         else:
             print("Options:",
                   "\n--------\n",
+                  "get: Make Dataset\n",
                   "train: Train model\n",
+                  "transform: Transfrom raw data in features using tsfresh\n",
                   "test: Test model performance\n",
                   "load: Load trained model and processed datasets\n",
                   "save: Save trained model and processed datasets\n",
                   "exit: Exit program")
             action = input("What do you want to do? ").lower()
 
-        if action == 'train':
+        if action == 'get':
+            print("Obtaining raw data from servers...")
+            start = timer()
+            for t in table_names:
+                make_dataset(t)
+                print("Table '{}' was downloaded".format(t))
+            end = timer()
+            print("Finished. ({0:.1f} sec elapsed)\n".format(end - start))
+        elif action == 'train':
             if x_train.empty or y_train.empty:
                 print("Generating train and test set from raw data...")
                 start = timer()
-                x_train, x_test, y_train, y_test = _split_raw()
+                x_train, x_test, y_train, y_test = transform_raw(feats, report_warning=report_warning, frac=0.99)
+                # TODO Test that the returned types are dataframes for X and series for Y.
+                # TODO Test the shapes are correct
+                # TODO Test the Y has different values
                 end = timer()
-                print("Finished. Train and test set loaded in memory. ({0:.1f} sec elapsed)".format(end - start))
+                print("Finished. Train and test set loaded in memory. ({0:.1f} sec elapsed)\n".format(end - start))
 
             # Obtain best model
             print("Training model...")
             start = timer()
-            model = train_model(x_train, y_train, grid_params)
+            model = train_model(x_train, y_train, grid_params, scoring, cv=cv)
             end = timer()
-            print("Finished. The trained model was loaded in memory. ({0:.1f} sec elapsed)".format(end - start))
+            print("Finished. The trained model was loaded in memory. ({0:.1f} sec elapsed)\n".format(end - start))
+        elif action == 'transform':
+            print("Generating train and test set from raw data...")
+            start = timer()
+            x_train, x_test, y_train, y_test = transform_raw(feats, report_warning=report_warning, frac=0.20)
+            end = timer()
+            print("Finished. Train and test set loaded in memory. ({0:.1f} sec elapsed)\n".format(end - start))
         elif action == 'load':
             print("Loading trained model and datasets...")
             start = timer()
-            model = _load_model()
-            x_train, x_test, y_train, y_test = _load_processed()
+            model = load_model()
+            x_train, y_train = load_processed(name='train')
+            x_test, y_test = load_processed(name='test')
+            # TODO Test that the returned types are dataframes for X and series for Y.
+            # TODO Test the shapes are correct
+            # TODO Test the Y has different values
             end = timer()
-            print("Finished. ({0:.1f} sec elapsed)".format(end - start))
-        elif model is None and (action == 'test' or action == 'save'):
-            print("There is no model loaded. Use 'train' or 'load' options.")
+            print("Finished. ({0:.1f} sec elapsed)\n".format(end - start))
         elif action == 'test':
-            print("Testing trained model...")
-            start = timer()
-            test_model(model, x_train, x_test, y_train, y_test)
-            end = timer()
-            print("Finished. ({0:.1f} sec elapsed)".format(end - start))
+            if model is None:
+                print("There is no model loaded. Use 'train' or 'load' options.")
+            else:
+                print("Testing trained model...")
+                start = timer()
+                test_model(model, x_test, y_test)
+                end = timer()
+                print("Finished. ({0:.1f} sec elapsed)\n".format(end - start))
         elif action == 'save':
-            print("Saving trained model and datasets...")
-            start = timer()
-            # Save trained model
-            _save_model(model)
-            # Save train and test sets in data/processed/
-            _save_processed(x_train, x_test, y_train, y_test)
-            end = timer()
-            print("Finished. ({0:.1f} sec elapsed)".format(end - start))
+            if x_train.empty or y_train.empty or x_test.empty or y_test.empty:
+                print("There is nothing to save. Use 'train' or 'load' options.")
+            else:
+                start = timer()
+                if model is None:
+                    print("No model detected. Saving only datasets...")
+                else:
+                    print("Saving trained model and datasets...")
+                    save_model(model)
+                    if model.best_estimator_.named_steps['clf'].tree_.node_count > nodes_warning:
+                        print('The tree is too big. The structure WAS NOT exported to a figure')
+                    else:
+                        print('Saving decision tree in reports/tree.png....')
+                        plot_tree(model.best_estimator_.named_steps['clf'], x_train.columns)
+
+                    print('Saving confusion matrix in reports/...')
+                    plot_confusion_matrix(model.best_estimator_, x_train, y_train, title=title)
+                    print('Saving plotted curves in reports/...')
+                    plot_precision_recall_curve(model.best_estimator_, x_train, y_train, title=title)
+                    plot_roc_curve(model.best_estimator_, x_train, y_train, x_test, y_test, title=title)
+                    plot_learning_curve(model.best_estimator_, x_train, y_train, scoring, cv=cv, title=title)
+                # Save train and test sets in data/processed/
+                save_processed(x_train, y_train, name='train')
+                save_processed(x_test, y_test, name='test')
+                end = timer()
+                print("Finished. ({0:.1f} sec elapsed)\n".format(end - start))
         elif action == 'exit':
             g_end = timer()
-            print("Total elapsed time: {0:.1f} seconds".format(g_end-g_start))
+            print("Total elapsed time: {0:.1f} seconds\n".format(g_end-g_start))
             break
         else:
             print("Unrecognized option '{}'".format(action))